pyproject.toml
[project]
name = "makatea"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mypy>=1.15.0",
    "pydantic>=2.11.3",
    "python-dateutil>=2.8.2",
    "ruff>=0.11.6",
    "textual>=3.1.0",
    "textual-dev>=1.7.0",
    "textual-plotext>=1.0.1",
    "tinydb>=4.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
